# docs/swagger.yaml
basePath: /
definitions:
...
  models.BancoOrigenList:
    properties:
      banco:
        type: string
      id:
        type: integer
      moneda:
        type: string
      nombre:
        type: string
    type: object
  models.CuentasBanco:
    properties:
      bancos_cliente:
        items:
          $ref: '#/definitions/models.BancoOrigenList'
        type: array
      metodos_pago_disponibles:
        items:
          $ref: '#/definitions/models.FormaPagoList'
        type: array
      tasa_cambio:
        type: number
    type: object
  models.ErrorResponse:
    properties:
      error: {}
    type: object
...
  models.ForgotPasswordRequest:
    properties:
      username:
        maxLength: 30
        minLength: 7
        type: string
    required:
    - username
    type: object
  models.ForgotPasswordSend:
    properties:
      password:
        maxLength: 30
        minLength: 7
        type: string
      token:
        type: string
      username:
        maxLength: 30
        minLength: 7
        type: string
    required:
    - password
    - token
    - username
    type: object
  models.FormaPagoList:
    properties:
      banco:
        type: string
      detalle: {}
      id:
        type: integer
      metodo_pago:
        type: string
      moneda:
        type: string
      nombre:
        type: string
    type: object
...
  models.SuccessResponse:
    properties:
      notice:
        type: string
      record: {}
    type: object
...
  models.UserChangePassword:
    properties:
      password:
        maxLength: 30
        minLength: 7
        type: string
      username:
        maxLength: 30
        minLength: 7
        type: string
    required:
    - password
    - username
    type: object
  models.UserCredentials:
    properties:
      change_passwd:
        type: boolean
    type: object
  models.UserDeviceInfo:
    properties:
      manufacturer:
        maxLength: 30
        minLength: 7
        type: string
      model:
        maxLength: 30
        minLength: 7
        type: string
      platform:
        maxLength: 30
        minLength: 6
        type: string
      screen_size:
        maxLength: 30
        minLength: 7
        type: string
      system:
        maxLength: 30
        minLength: 7
        type: string
      user_agent:
        maxLength: 200
        minLength: 7
        type: string
    type: object
  models.UserProfile:
    properties:
      correo:
        items:
          type: string
        type: array
      direccion:
        type: string
      id:
        type: string
      nombre:
        type: string
      telefono:
        items:
          type: string
        type: array
      username:
        type: string
    type: object
  models.UserRequest:
    properties:
      device_info:
        $ref: '#/definitions/models.UserDeviceInfo'
      password:
        maxLength: 30
        minLength: 7
        type: string
      remember:
        type: string
      username:
        maxLength: 30
        minLength: 7
        type: string
    required:
    - password
    - remember
    - username
    type: object
  models.UserResponse:
    properties:
      credentials:
        $ref: '#/definitions/models.UserCredentials'
      profile:
        $ref: '#/definitions/models.UserProfile'
      tokens:
        $ref: '#/definitions/models.UserToken'
    type: object
  models.UserToken:
    properties:
      auth:
        type: string
      auth_expires_at:
        type: string
      refresh:
        type: string
      refresh_expires_at:
        type: string
    type: object
host: 127.0.0.1:7003
info:
  contact:
    email: juan9acosta@gmail.com
    name: Juan Acosta
    url: https://www.linkedin.com/in/juan-m-acosta-f-54219758/
  description: service in Go using Gin framework
  title: MiCuenta Service API
  version: "1.0"
paths:
  /auth/change-password:
    post:
      consumes:
      - application/json
      description: Updates the user's password after validating the request body
      parameters:
      - description: Access Token
        in: header
        name: x-access-token
        required: true
        type: string
      - description: Password Change Request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserChangePassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Invalid Request or Incorrect Data
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Change user password
      tags:
      - Authentication
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Sends a password reset email to the user after validating the request
      parameters:
      - description: User email for password reset
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Invalid Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: User Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Request password reset
      tags:
      - Authentication
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user by validating credentials and returning a
        session token
      parameters:
      - description: Access Token
        in: header
        name: x-access-token
        type: string
      - description: User Login Data
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/models.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                record:
                  $ref: '#/definitions/models.UserResponse'
              type: object
        "400":
          description: Invalid Request or Already Logged In
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Authenticate a user
      tags:
      - Authentication
  /auth/logout:
    get:
      consumes:
      - application/json
      description: Logs out the authenticated user by clearing session data
      parameters:
      - description: Access Token
        in: header
        name: x-access-token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
      summary: Logout a user
      tags:
      - Authentication
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Refreshes the user's session token if valid and returns a new one
      parameters:
      - description: Access Token
        in: header
        name: x-access-token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                record:
                  $ref: '#/definitions/models.UserToken'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Refresh authentication token
      tags:
      - Authentication
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Processes a password reset request and sends recovery instructions
        to the user
      parameters:
      - description: User email for password reset
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.ForgotPasswordSend'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset instructions sent
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Invalid Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: User Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Send password reset instructions
      tags:
      - Authentication
  /banco/formas-de-pago:
    get:
      consumes:
      - application/json
      description: devuelve un listado de las formas de pago disponibles para el cliente
      parameters:
      - description: Access Token
        in: header
        name: x-access-token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                record:
                  $ref: '#/definitions/models.CuentasBanco'
              type: object
        "400":
          description: Invalid Request or Incorrect Data
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: listado formas de pago
      tags:
      - Banco
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"